move test into its own test_verify_disagg02
